renamed simple_neural_network directory from SimpleNeuralNetwork
