+++ drevalpy/models/simple_neural_network/utils.py
