+++ setup.py
@@ -39,7 +39,7 @@ setup(
     package_data={
         "": [
             "models/baselines/hyperparameters.yaml",
-            "models/simpleNeuralNetwork/hyperparameters.yaml",
+            "models/simple_neural_network/hyperparameters.yaml",
             "visualization/style_utils/favicon.png",
             "visualization/style_utils/index_layout.html",
             "visualization/style_utils/LCO.png",
